requirements.txt
loguru==0.7.2
openai-whisper==20231117
httpx==0.27.0
